10.13밤pull

+++ src/main/java/com/project/mohe/controller/AdminController.java
@@ -2,6 +2,9 @@ package com.project.mohe.controller;
 
 import java.util.HashMap;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -27,7 +30,7 @@ public class AdminController {
 	@Autowired
 	private PagingService pagingService;
 	@Autowired
-	private DonationService domainService;
+	private DonationService donationService;
 	@Autowired
 	private PopupService popupService;
 	@Autowired
@@ -35,29 +38,59 @@ public class AdminController {
 	
 	// 관리자 화면 자동이동을 위한 메소드
 	@RequestMapping("{step}.do")
-	public String viewPage(@PathVariable String step) {
+	public String viewPage(@PathVariable String step,HttpServletRequest request) {
+		// 로그인 하지않은 사람이 접근할 수 없도록
+		if(request.getSession().getAttribute("adm_no") == null) return "/admin/adminLogin";
 		return "/admin/"+step;
 	}
 	// 관리자 목록 띄우기
 	@RequestMapping("adminList.do")
-	public void getAdminList(PagingVO vo,HashMap map,Model model) {	
+	public String getAdminList(PagingVO vo,Model model,HttpServletRequest request) {	
+		// 로그인 하지않은 사람이 접근할 수 없도록
+		if(request.getSession().getAttribute("adm_no") == null) return "/admin/adminLogin";
 		// 페이징을 위한 테이블 행값 받아오기
-		vo.setTotalRecCount(adminService.getAllcnt().getTotalRecCount());
+		vo.setTotalRecCount(adminService.getAllcnt(vo).getTotalRecCount());
 		// 페이징 처리 
 		vo = pagingService.doPaging(vo);
 		// 페이지값 저장하기
 		model.addAttribute("page",vo);
 		// 페이징을 토대로한 리스트 목록 불러오기
-		model.addAttribute("adminList",adminService.getAdminList(vo,map));
+		model.addAttribute("adminList",adminService.getAdminList(vo));
+		return "/admin/adminList";
 	}
 	// 관리자 상세 목록 띄우기
 	@RequestMapping("adminDetail.do")
-	public void getAdminDetail(AdminVO vo,Model model) {
+	public String getAdminDetail(AdminVO vo,Model model,HttpServletRequest request) {
+		// 로그인 하지않은 사람이 접근할 수 없도록
+		if(request.getSession().getAttribute("adm_no") == null) return "/admin/adminLogin";
 		model.addAttribute("admin",adminService.getAdminDetail(vo));
+		return "/admin/adminDetail";
+	}
+	// 관리자 로그인 확인
+	@RequestMapping("getAdminLogin.do")
+	public String adminLogin(AdminVO vo,HttpServletRequest request) {
+		// 아이디와 비밀번호가 맞아서 값이 비어있지 않다면
+		if(adminService.adminLogin(vo) != null) {
+			HttpSession session = request.getSession();
+			session.setAttribute("adm_no", vo.getAdm_no());
+			return "redirect:/admin/adminMain.do"; 
+		}
+		// 아이디와 비밀번호 값이 다르면 false값 전달
+		return "redirect:/admin/adminLogin.do?result=false"; 
+	}
+	// 관리자 로그아웃
+	@RequestMapping("adminLogout.do")
+	public String adminLogout(HttpServletRequest request) {
+		if(request.getSession().getAttribute("adm_no") != null) {
+			request.getSession().invalidate();
+		}
+		return "redirect:/admin/adminLogin.do";
 	}
 	// 회원 목록 띄우기
 	@RequestMapping("adUserList.do")
-	public void getUserList(PagingVO vo,HashMap map,Model model) {	
+	public String getUserList(PagingVO vo,HashMap map,Model model,HttpServletRequest request) {	
+		// 로그인 하지않은 사람이 접근할 수 없도록
+		if(request.getSession().getAttribute("adm_no") == null) return "/admin/adminLogin";
 		// 페이징을 위한 테이블 행값 받아오기
 		vo.setTotalRecCount(userInfoService.getAllcnt().getTotalRecCount());
 		// 페이징 처리 
@@ -66,95 +99,135 @@ public class AdminController {
 		model.addAttribute("page",vo);
 		// 페이징을 토대로한 리스트 목록 불러오기
 		model.addAttribute("userList",adminService.adGetUserList(vo,map));
+		return "/admin/adUserList";
 	}
 	// 회원 상세 페이지
 	@RequestMapping("adUserDetail.do")
-	public void adUserDetail(HashMap map,Model model,UserInfoVO vo) {	
+	public String adUserDetail(HashMap map,Model model,UserInfoVO vo,HttpServletRequest request) {
+		// 로그인 하지않은 사람이 접근할 수 없도록
+		if(request.getSession().getAttribute("adm_no") == null) return "/admin/adminLogin";	
 		model.addAttribute("userDetail",adminService.adUserDetail(vo));
+		return "/admin/adUserDetail";
 	}
 	// 이벤트 목록 띄우기
 	@RequestMapping("adEventList.do")
-	public void getEventList(HashMap map,Model model) {	
+	public String getEventList(HashMap map,Model model,HttpServletRequest request) {	
+		// 로그인 하지않은 사람이 접근할 수 없도록
+		if(request.getSession().getAttribute("adm_no") == null) return "/admin/adminLogin";
 		model.addAttribute("eventList",adminService.adGetEventList(map));
+		return "/admin/adEventList";
 	}
 	// 승인된 펀딩 목록
 	@RequestMapping("adFdList.do")
-	public void getFundingList(HashMap map,Model model) {
+	public String getFundingList(HashMap map,Model model,HttpServletRequest request) {
+		// 로그인 하지않은 사람이 접근할 수 없도록
+		if(request.getSession().getAttribute("adm_no") == null) return "/admin/adminLogin";
 		model.addAttribute("fdList",adminService.adGetFdList(map)); 
+		return "/admin/adFdList";
 	}
 	// 승인되지 않은 펀딩 목록
 	@RequestMapping("adFdApproval.do")
-	public void getFdApprovalList(HashMap map,Model model) {
+	public String getFdApprovalList(HashMap map,Model model,HttpServletRequest request) {
+		// 로그인 하지않은 사람이 접근할 수 없도록
+		if(request.getSession().getAttribute("adm_no") == null) return "/admin/adminLogin";
 		model.addAttribute("fdApproval",adminService.getFdApprovalList(map)); 
+		return "/admin/adFdApproval";
 	}
 	// 펀딩 승인,비승인 업데이트
 	@RequestMapping("judgBsUpdate.do")
-	public String judgBsUpdate(BongsaVO vo,Model model) {
+	public String judgBsUpdate(BongsaVO vo,Model model,HttpServletRequest request) {
+		// 로그인 하지않은 사람이 접근할 수 없도록
+		if(request.getSession().getAttribute("adm_no") == null) return "/admin/adminLogin";
 		//bs_judg 변수를 이용해 승인인지, 비승인인지 service에서 판단후 업데이트 
 		adminService.judgBsUpdate(vo);
 		return "redirect:/admin/adFdApproval.do";
 	}
 	// 승인된 봉사 목록
 	@RequestMapping("adVtList.do")
-	public void getBsList(HashMap map,Model model) {
+	public String getBsList(HashMap map,Model model,HttpServletRequest request) {
+		// 로그인 하지않은 사람이 접근할 수 없도록
+		if(request.getSession().getAttribute("adm_no") == null) return "/admin/adminLogin";
 		model.addAttribute("bsList",adminService.getBsList(map)); 
+		return "/admin/adVtList";
 	}
 	// 승인되지 않은 봉사 목록
 	@RequestMapping("adVtApproval.do")
-	public void getBsApprovalList(HashMap map,Model model) {
+	public String getBsApprovalList(HashMap map,Model model,HttpServletRequest request) {
+		// 로그인 하지않은 사람이 접근할 수 없도록
+		if(request.getSession().getAttribute("adm_no") == null) return "/admin/adminLogin";
 		model.addAttribute("bsApproval",adminService.getBsApprovalList(map)); 
+		return "/admin/adVtApproval";
 	}
 	// 펀딩 승인,비승인 업데이트
 	@RequestMapping("judgFdUpdate.do")
-	public String judgFdUpdate(Funding_pjVO vo,Model model) {
+	public String judgFdUpdate(Funding_pjVO vo,Model model,HttpServletRequest request) {
+		// 로그인 하지않은 사람이 접근할 수 없도록
+		if(request.getSession().getAttribute("adm_no") == null) return "/admin/adminLogin";
 		//fd_judg 변수를 이용해 승인인지, 비승인인지 service에서 판단후 업데이트 
 		adminService.judgFdUpdate(vo);
 		return "redirect:/admin/adFdApproval.do";
 	}
 	// 파트너쉽 목록
 	@RequestMapping("adPartnerList.do")
-	public void getPartnerList(HashMap map,Model model) {
+	public String getPartnerList(HashMap map,Model model,HttpServletRequest request) {
+		// 로그인 하지않은 사람이 접근할 수 없도록
+		if(request.getSession().getAttribute("adm_no") == null) return "/admin/adminLogin";
 		model.addAttribute("partner",adminService.getPartnerList(map)); 
+		return "/admin/adPartnerList";
 	}
 	// 리뷰 목록
 	@RequestMapping("adReviewList.do")
-	public void getReviewList(HashMap map,Model model) {
+	public String getReviewList(HashMap map,Model model,HttpServletRequest request) {
+		// 로그인 하지않은 사람이 접근할 수 없도록
+		if(request.getSession().getAttribute("adm_no") == null) return "/admin/adminLogin";
 		model.addAttribute("review",adminService.getReviewList(map)); 
+		return "/admin/adReviewList";
 	}
 	// 기부 목록
 	@RequestMapping("adDonationList.do")
-	public void getDonationList(PagingVO vo,HashMap map,Model model) {
+	public String getDonationList(PagingVO vo,Model model,HttpServletRequest request) {
+		// 로그인 하지않은 사람이 접근할 수 없도록
+		if(request.getSession().getAttribute("adm_no") == null) return "/admin/adminLogin";
 		// 페이징을 위한 테이블 행값 받아오기
-		vo.setTotalRecCount(domainService.getAllcnt().getTotalRecCount());
+		vo.setTotalRecCount(donationService.getAllcnt(vo).getTotalRecCount());
 		// 페이징 처리 
 		vo = pagingService.doPaging(vo);
 		// 페이지값 저장하기
 		model.addAttribute("page",vo);
 		// 페이징을 토대로한 리스트 목록 불러오기
-		model.addAttribute("donation",adminService.getDonationList(vo,map)); 
+		model.addAttribute("donation",adminService.getDonationList(vo)); 
+		return "/admin/adDonationList";
 	}
 	// 공지 리스트
 	@RequestMapping("adNotice.do")
-	public void getNoticeList(HashMap map,Model model) {
+	public String getNoticeList(HashMap map,Model model,HttpServletRequest request) {
+		// 로그인 하지않은 사람이 접근할 수 없도록
+		if(request.getSession().getAttribute("adm_no") == null) return "/admin/adminLogin";
 		model.addAttribute("notice",adminService.getNoticeList(map)); 
+		return "/admin/adNotice";
 	}
 	// 팝업 리스트
 	@RequestMapping("adPopList.do")
-	public void getPopupList(PagingVO vo,HashMap map,Model model) {
+	public String getPopupList(PagingVO vo,Model model,HttpServletRequest request) {
+		// 로그인 하지않은 사람이 접근할 수 없도록
+		if(request.getSession().getAttribute("adm_no") == null) return "/admin/adminLogin";
 		// 페이징을 위한 테이블 행값 받아오기
-		vo.setTotalRecCount(popupService.getAllcnt().getTotalRecCount());
+		vo.setTotalRecCount(popupService.getAllcnt(vo).getTotalRecCount());
 		// 페이징 처리 
 		vo = pagingService.doPaging(vo);
 		// 페이지값 저장하기
 		model.addAttribute("page",vo);
 		// 페이징을 토대로한 리스트 목록 불러오기
-		model.addAttribute("popup",adminService.getPopupList(vo,map)); 
+		model.addAttribute("popup",adminService.getPopupList(vo)); 
+		return "/admin/adPopList";
 	}
 
 	
 	// 관리자 사용불가 전환하기
 	@RequestMapping("adminEnd.do")
-	public String deleteAdmin(AdminVO vo) {
+	public String deleteAdmin(AdminVO vo,HttpServletRequest request) {
+		// 로그인 하지않은 사람이 접근할 수 없도록
+		if(request.getSession().getAttribute("adm_no") == null) return "/admin/adminLogin";
 		// 퇴사일이 시스데이트로 업데이트됨
 		adminService.deleteAdmin(vo);
 		return "redirect:/admin/adminList.do";
@@ -162,14 +235,19 @@ public class AdminController {
 	
 	// 관리자 수정페이지 이동하기
 	@RequestMapping("adminUpdateInfo.do")
-	public void adminUpdateInfo(AdminVO vo,Model model) {
+	public String adminUpdateInfo(AdminVO vo,Model model,HttpServletRequest request) {
+		// 로그인 하지않은 사람이 접근할 수 없도록
+		if(request.getSession().getAttribute("adm_no") == null) return "/admin/adminLogin";
 		// 퇴사일이 시스데이트로 업데이트됨
 		model.addAttribute("admin",adminService.getAdminDetail(vo));
+		return "/admin/adminUpdateInfo";
 	}
 	
 	// 관리자정보 수정하기
 	@RequestMapping("updateAdmin.do")
-	public String updateAdmin(AdminVO vo) {
+	public String updateAdmin(AdminVO vo,HttpServletRequest request) {
+		// 로그인 하지않은 사람이 접근할 수 없도록
+		if(request.getSession().getAttribute("adm_no") == null) return "/admin/adminLogin";
 		adminService.updateAdmin(vo);
 		return "redirect:/admin/adminDetail.do?adm_no="+vo.getAdm_no();
 	}
